Removes v0 healthcare service transformers (#25024)
* Removes v0 healthcare service transformers


* Removes unit tests for v0 healhcare service transformers


* Renames v2 transformers file


---------
