+++ src/applications/vaos/services/healthcare-service/transformers.js
@@ -3,44 +3,34 @@
  */
 
 /**
- * Transforms
- * /vaos/v0/facilities/{facilityId}/clinics
- * to
- * /HealthcareService
+ * Transforms a clinic from the VAOS service into our HealthCareService format
  *
- * @param {string} facilityId
- * @param {VARClinic} clinic Clinic object to transform to HealthcareService object
+ * @param {VAOSClinic} clinic Clinic object to transform to HealthcareService object
  *
- * @returns {HealthCareService} A FHealthCareService resource
+ * @returns {HealthCareService} A HealthCareService resource
  */
-export function transformAvailableClinic(facilityId, typeOfCareId, clinic) {
+export function transformClinicV2(clinic) {
   return {
     // ID: VA facility site code and clinic id
-    id: `${clinic.siteCode}_${clinic.clinicId}`,
+    id: `${clinic.vistaSite}_${clinic.id}`,
     // ID of physical VA Facility (sta6id) where the clinic is located
-    stationId: facilityId,
+    stationId: clinic.stationId,
     // Name of VA facility where clinic is located
-    stationName: clinic.institutionName,
+    stationName: clinic.stationName,
     // Description of service as presented to a consumer while searching
-    serviceName: clinic.clinicFriendlyLocationName || clinic.clinicName,
-    // String value to allow patient direct scheduling  Y/N
-    patientDirectScheduling: clinic.directSchedulingFlag,
+    serviceName: clinic.serviceName,
+    // Boolean to allow direct scheduling by patient
+    patientDirectScheduling: clinic.patientDirectScheduling,
   };
 }
 
 /**
- * Transforms
- * /vaos/v0/facilities/{facilityId}/clinics
- * to
- * /HealthcareService
+ * Transforms a list of clinics from the VAOS service into our HealthCareService format
  *
- * @param {string} facilityId
- * @param {Array<VARClinic>} clinics Array of clinic objects to transform to HealthcareService objects
+ * @param {Array<VAOSClinic>} clinics Array of clinic objects to transform to HealthcareService objects
  *
  * @returns {Array<HealthCareService>} An array of FHIR HealthcareService objects
  */
-export function transformAvailableClinics(facilityId, typeOfCareId, clinics) {
-  return clinics.map(clinic =>
-    transformAvailableClinic(facilityId, typeOfCareId, clinic),
-  );
+export function transformClinicsV2(clinics) {
+  return clinics.map(clinic => transformClinicV2(clinic));
 }
